Fix back button on ec2 instances page
Essentially, what was happening previously was that after terminating a sid box, if you pressed the browser back button, it would return to the HTML page, and the sniffles-destroy job would still be loaded in the iframe. This would cause the `onload` to be triggered, and we would assume that you had just terminated a job, and redirect the main window to sniffles-destroy again.

This fix dynamically adds the <iframe> after the page loads, so when you press the browser's _back button_, there is no iframe initially. And it will be re-initialised on `about:blank`

As always there are multiple ways to fix this, probably simpler would be to add a global variable that only becomes set when you press a "terminate" button. Then only redirect if both variables are true. But I wanted to do this without adding extra state, also I wanted to see wth was going on with the back button